Temporary fix for images without proper nether patches.
If the nether patches are not in the kernel, the rule
that was commented out, will stop all outgoing network traffic.
This should not be the case thanks to the queue-bypass
parameter to iptables, but it seems to fail anyway.

Since the kernel patches are not yet merged, nether is
useless anyway. This will fix any issues until this changes.

Change-Id: Ic6c6876a62588f76d0f7e4105d2866320474149f